+++ doc/source/whats_new.rst
@@ -45,6 +45,8 @@ Changelog
    
    - Add `ico` and `oct` source space creation in native Python by `Eric Larson`_
 
+   - Add interactive rejection of bad trials in `plot_epochs` by `Denis Engemann`_
+
 API
 ~~~